Venn Diagrams! (in R!)
It seems not many are familiar with drawing Venn Diagrams on R and prefer websites (e.g. Venny 2.1.0 which has a max of 4 circles).

If you want to learn, I've uploaded code that I've used in the past, and some data from which to plot them if you want to play. 
Essentially its drawing a Venn diagram to show the overlap in client proteins between various chaperone proteins, using the package VennDiagram and gridExtra. Very easy stuff!

Hope it helps,
Becki